+++ src/processing/matchISLES2018.m
@@ -15,8 +15,6 @@
 % - NCCT is int32, CBF and CBV are uint16, and MTT and Tmax are double.
 % - Each image has a different ideal display range using imshow.
 % 
-% Requires at least MATLAB 2020a (for exportgraphics)
-% 
 %   Kyle See 10/24/2023
 %   Smart Medical Informatics Learning and Evaluation (SMILE) Laboratory
 %   Biomedical Engineering
@@ -33,12 +31,11 @@
 %#########################################
 clc; clear; close all; warning off;
 % Input path MUST contain the unzipped TRAINING and TESTING folders.
-inputPath = 'D:\Desktop Files\Dropbox (UFL)\Quick Coding Scripts\Testing MAGIC pipeline\isles_deid';
-partitionPath = 'D:\Desktop Files\Dropbox (UFL)\Quick Coding Scripts\Testing MAGIC pipeline\isles_partition';
+inputPath = 'C:\Users\kylebsee\Dropbox (UFL)\Quick Coding Scripts\Testing MAGIC pipeline\isles_deid';
+partitionPath = 'C:\Users\kylebsee\Dropbox (UFL)\Quick Coding Scripts\Testing MAGIC pipeline\isles_partition';
 %#########################################
 
 %% Initialization
-
 fprintf("Starting...matchISLES2018.m\n")
 fprintf("------------------------------------------------------------------\n")
 
@@ -55,15 +52,74 @@ if ~exist(rCBFPath,'dir'), mkdir(rCBFPath); end
 if ~exist(MTTPath,'dir'), mkdir(MTTPath); end
 if ~exist(NCCTsavePath,'dir'), mkdir(NCCTsavePath); end
 
-slice_count = [];
+% % Training Loop - hardcoded to the downloaded data for ease
+% for i = 1:94
+%     % Read case number in
+%     caseFile = strcat('case_',num2str(i));
+% 
+%     % Set filepath for case
+%     baseFolder = fullfile(inputPath,'TRAINING',caseFile);
+% 
+%     % Get directory list
+%     % - In order, CT, DWI, CBF, CBV, MTT, Tmax, OT
+%     baseDir = dir(baseFolder);
+% 
+%     % Remove .and ..
+%     baseDir(1:2) = [];
+% 
+%     % Read in images based on positions
+%     ct = niftiread(fullfile(baseDir(1).folder,baseDir(1).name,strcat(baseDir(1).name,'.nii')));
+%     cbf = niftiread(fullfile(baseDir(3).folder,baseDir(3).name,strcat(baseDir(3).name,'.nii')));
+%     cbv = niftiread(fullfile(baseDir(4).folder,baseDir(4).name,strcat(baseDir(4).name,'.nii')));
+%     mtt = niftiread(fullfile(baseDir(5).folder,baseDir(5).name,strcat(baseDir(5).name,'.nii')));
+%     tmax = niftiread(fullfile(baseDir(6).folder,baseDir(6).name,strcat(baseDir(6).name,'.nii')));
+% 
+%     % Cycle through all of the slices
+%     num_slices = size(cbf,3);
+%     for j = 1:num_slices
+%         if i < 10
+%             saveName = strcat('10000',num2str(i),'_0',num2str(j),'.png');
+%         else
+%             saveName = strcat('1000',num2str(i),'_',num2str(j),'.png');
+%         end
+% 
+%         % Copy image twice to simulate pseudo-RGB input
+%         savePathNCCT = fullfile(partitionPath,'NCCT',saveName);
+%         sliceNCCT = imrotate(ct(:,:,j),90);
+%         sliceNCCT(:,:,2) = imrotate(ct(:,:,j),90);
+%         sliceNCCT(:,:,3) = imrotate(ct(:,:,j),90);
+%         sliceNCCT = uint8(sliceNCCT);
+%         imwrite(sliceNCCT,savePathNCCT)
+% 
+%         savePathCBF = fullfile(partitionPath,'rCBF',saveName);
+%         sliceCBF = uint8(imrotate(cbf(:,:,j),90));
+%         sliceCBF = imadjust(sliceCBF);
+%         imwrite(sliceCBF,savePathCBF)
+% 
+%         savePathCBV = fullfile(partitionPath,'rCBV',saveName);
+%         sliceCBV = uint8(imrotate(cbv(:,:,j),90));
+%         sliceCBV = imadjust(sliceCBV);
+%         imwrite(sliceCBV,savePathCBV)
+% 
+%         savePathMTT = fullfile(partitionPath,'MTT',saveName);
+%         sliceMTT = uint8(imrotate(mtt(:,:,j),90));
+%         sliceMTT = imadjust(sliceMTT);
+%         imwrite(sliceMTT,savePathMTT)
+% 
+%         savePathTTP = fullfile(partitionPath,'TTP',saveName);
+%         sliceTTP = uint8(imrotate(tmax(:,:,j),90));
+%         sliceTTP = imadjust(sliceTTP);
+%         imwrite(sliceTTP,savePathTTP)
+%     end
+% end
 
-% Training Loop - hardcoded to the downloaded data for ease
-for i = 1:94
+% Testing Loop - hardcoded to the downloaded data for ease
+for i = 1:62 % 63 is a file called .DS_Store
     % Read case number in
     caseFile = strcat('case_',num2str(i));
     
     % Set filepath for case
-    baseFolder = fullfile(inputPath,'TRAINING',caseFile);
+    baseFolder = fullfile(inputPath,'TESTING',caseFile);
     
     % Get directory list
     % - In order, CT, DWI, CBF, CBV, MTT, Tmax, OT
@@ -79,21 +135,16 @@ for i = 1:94
     mtt = niftiread(fullfile(baseDir(5).folder,baseDir(5).name,strcat(baseDir(5).name,'.nii')));
     tmax = niftiread(fullfile(baseDir(6).folder,baseDir(6).name,strcat(baseDir(6).name,'.nii')));
 
-    % CT looks best with 100
-    % CBF looks best with 1000
-    % CBV looks best with 100
-    % MTT looks best with 20
-    % Tmax looks best with 20
-    
+    % Cycle through all of the slices
     num_slices = size(cbf,3);
-    
     for j = 1:num_slices
         if i < 10
-            saveName = strcat('10000',num2str(i),'_0',num2str(j),'.png');
+            saveName = strcat('20000',num2str(i),'_0',num2str(j),'.png');
         else
-            saveName = strcat('1000',num2str(i),'_',num2str(j),'.png');
+            saveName = strcat('2000',num2str(i),'_',num2str(j),'.png');
         end
         
+        % Copy image twice to simulate pseudo-RGB input
         savePathNCCT = fullfile(partitionPath,'NCCT',saveName);
         sliceNCCT = imrotate(ct(:,:,j),90);
         sliceNCCT(:,:,2) = imrotate(ct(:,:,j),90);
@@ -103,41 +154,25 @@ for i = 1:94
         
         savePathCBF = fullfile(partitionPath,'rCBF',saveName);
         sliceCBF = uint8(imrotate(cbf(:,:,j),90));
-        figure('Visible','off');
-        imshow(sliceCBF, [0 max(sliceCBF(:))])
-        saveas(gcf,savePathCBF,'png')
-        close;
+        sliceCBF = imadjust(sliceCBF);
+        imwrite(sliceCBF,savePathCBF)
         
         savePathCBV = fullfile(partitionPath,'rCBV',saveName);
         sliceCBV = uint8(imrotate(cbv(:,:,j),90));
-        figure('Visible','off');
-        imshow(sliceCBV,[0 max(sliceCBV(:))])
-        saveas(gcf,savePathCBV,'png')
-        close;
+        sliceCBV = imadjust(sliceCBV);
+        imwrite(sliceCBV,savePathCBV)
         
         savePathMTT = fullfile(partitionPath,'MTT',saveName);
         sliceMTT = uint8(imrotate(mtt(:,:,j),90));
-        figure();
-        imshow(sliceMTT, [0 max(sliceMTT(:))])
-        saveas(gcf,savePathMTT,'png')
-        close;
+        sliceMTT = imadjust(sliceMTT);
+        imwrite(sliceMTT,savePathMTT)
         
         savePathTTP = fullfile(partitionPath,'TTP',saveName);
         sliceTTP = uint8(imrotate(tmax(:,:,j),90));
-        figure('Visible','off');
-        imshow(sliceTTP, [0 max(sliceTTP(:))])
-        saveas(gcf,savePathTTP,'png')
-        close;
-        
+        sliceTTP = imadjust(sliceTTP);
+        imwrite(sliceTTP,savePathTTP)
     end
-    
-%     slice_count = [slice_count num_slices];
-%     figure;
-%     for j = 1:num_slices
-%         subplot(2,4,j)
-%         imshow(imrotate(ct(:,:,j),90),[0 100])
-%     end
-%     close;
 end
+
 
 % end